pyproject.toml
[project]
name = "little-lemon"
version = "0.1.0"
description = ""
authors = [
    {name = "Akashdeep singh",email = "akashdeep.singh@sifars.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0) ; python_version >= \"3.12\" and python_version < \"4.0\"",
    "bcrypt (==4.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
colorlog = "^6.9.0"
alembic = "^1.16.4"
